Split Engie examples out into their own files for easier merging
